tests/wpt/tests/html/canvas/offscreen/drawing-rectangles-to-the-canvas/2d.clearRect.zero.html

<!DOCTYPE html>
<!-- DO NOT EDIT! This test has been generated by /html/canvas/tools/gentest.py. -->
<meta charset="UTF-8">
<title>OffscreenCanvas test: 2d.clearRect.zero</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/html/canvas/resources/canvas-tests.js"></script>

<h1>2d.clearRect.zero</h1>
<p class="desc">clearRect of zero pixels has no effect</p>


<script>
var t = async_test("clearRect of zero pixels has no effect");
var t_pass = t.done.bind(t);
var t_fail = t.step_func(function(reason) {
    throw reason;
});
t.step(function() {

  var canvas = new OffscreenCanvas(100, 50);
  var ctx = canvas.getContext('2d');

  ctx.fillStyle = '#0f0';
  ctx.fillRect(0, 0, 100, 50);
  ctx.clearRect(0, 0, 100, 0);
  ctx.clearRect(0, 0, 0, 50);
  ctx.clearRect(0, 0, 0, 0);
  _assertPixel(canvas, 50,25, 0,255,0,255);
  t.done();

});
</script>
